requirements.txt
# FastAPI and core dependencies
fastapi==0.117.1
uvicorn==0.37.0
pydantic==2.11.9
pydantic-core==2.33.2

# AI and ML dependencies
google-generativeai==0.8.3
Pillow==11.0.0
numpy==1.24.3

# Database dependencies
SQLAlchemy==2.0.43
psycopg[binary]==3.2.3

# HTTP and networking
httpx==0.27.2
aiohttp==3.10.10
python-multipart==0.0.12

# Validation and schema
jsonschema==4.17.3

# Environment and configuration
python-dotenv==1.1.1
PyYAML==6.0.3

# Development and testing
pytest==8.3.3
pytest-asyncio==0.23.6

# Additional utilities
typing-extensions==4.15.0
annotated-types==0.7.0

# AI Agent dependencies with LangGraph
langgraph==0.6.8
langchain==0.3.27
langchain-google-genai==2.1.12
langchain-core==0.3.77